+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Release Notes
 
-
 ## Platform Version 0.9.11 - UNRELEASED
 * Reconnect producer to SPU if network error. ([#770](https://github.com/infinyon/fluvio/issues/770))
 * Merge fluvio-protocol-{api,core,codec} crates into fluvio-protocol ([#1594](https://github.com/infinyon/fluvio/issues/1594))
+* Add `#[smarstream(flat_map)]` for flattening one record into many ([#1335](https://github.com/infinyon/fluvio/issues/1335))
 
 ## Platform Version 0.9.10 - 2021-10-07
 * Improve error handling for socket timeout ([#791](https://github.com/infinyon/fluvio/issues/791))